feat: class based views

+++ django/playground/store/views_function_based.py
@@ -2,83 +2,82 @@ from django.shortcuts import get_object_or_404
 from django.db.models.aggregates import Count
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
-from rest_framework.views import APIView
 from rest_framework import status
 from .models import Product, Collection
 from .serializers import ProductSerializer, CollectionSerializer
 
 
-class ProductList(APIView):
-    def get(self, request):
+@api_view(['GET', 'POST'])
+def product_list(request):
+    if request.method == 'GET':
         queryset = Product.objects.select_related('collection').all()
         serializer = ProductSerializer(queryset, many=True)
         return Response(serializer.data)
-
-    def post(self, request):
+    elif request.method == 'POST':
         serializer = ProductSerializer(data=request.data)
+        # if serializer.is_valid():
+        #     serializer.validated_data
+        #     return Response('ok')
+        # else:
+        #     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
         serializer.is_valid(raise_exception=True)
         serializer.save()
         return Response(serializer.data, status=status.HTTP_201_CREATED)
 
 
-class ProductDetail(APIView):
-    def get(self, request, id):
-        product = get_object_or_404(Product, pk=id)
+# @api_view()
+# def product_detail(request, id):
+#     try:
+#         product = Product.objects.get(pk=id)
+#         serializer = ProductSerializer(product)
+#         return Response(serializer.data)
+#     except Product.DoesNotExist:
+#         return Response(status=status.HTTP_404_NOT_FOUND)
+
+@api_view(['GET', 'PUT', 'DELETE'])
+def product_detail(request, id):
+    product = get_object_or_404(Product, pk=id)
+    if request.method == 'GET':
         serializer = ProductSerializer(product)
         return Response(serializer.data)
-
-    def put(self, request, id):
-        product = get_object_or_404(Product, pk=id)
+    elif request.method == 'PUT':
         serializer = ProductSerializer(product, data=request.data)
         serializer.is_valid(raise_exception=True)
         serializer.save()
         return Response(serializer.data)
-
-    def delete(self, request, id):
-        product = get_object_or_404(Product, pk=id)
+    elif request.method == 'DELETE':
         if product.orderitems.count() > 0:
-            return Response(
-                {
-                    'error': 'there is an order item with this product'
-                },
-                status=status.HTTP_405_METHOD_NOT_ALLOWED)
+            return Response({'error': 'there is an order item with this product'}, status=status.HTTP_405_METHOD_NOT_ALLOWED)
         product.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 
-class CollectionList(APIView):
-    def get(self, request):
+@api_view(['GET', 'POST'])
+def collection_list(request):
+    if request.method == 'GET':
         queryset = Collection.objects.all()
         serializer = CollectionSerializer(queryset, many=True)
         return Response(serializer.data)
-
-    def post(self, request):
+    elif request.method == 'POST':
         serializer = CollectionSerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
         serializer.save()
         return Response(serializer.data, status=status.HTTP_201_CREATED)
 
 
-class CollectionDetail(APIView):
-    def get(self, request, pk):
-        collection = get_object_or_404(Collection, pk=pk)
+@api_view(['GET', 'PUT', 'DELETE'])
+def collection_detail(request, pk):
+    collection = get_object_or_404(Collection, pk=pk)
+    if request.method == 'GET':
         serializer = CollectionSerializer(collection)
         return Response(serializer.data)
-
-    def put(self, request, pk):
-        collection = get_object_or_404(Collection, pk=pk)
+    elif request.method == 'PUT':
         serializer = CollectionSerializer(collection, data=request.data)
         serializer.is_valid(raise_exception=True)
         serializer.save()
         return Response(serializer.data)
-
-    def delete(self, request, pk):
-        collection = get_object_or_404(Collection, pk=pk)
+    elif request.method == 'DELETE':
         if collection.products.count() > 0:
-            return Response(
-                {
-                    'error': 'collection has products and can not be deleted'
-                },
-                status=status.HTTP_405_METHOD_NOT_ALLOWED)
+            return Response({'error': 'collection has products and can not be deleted'}, status=status.HTTP_405_METHOD_NOT_ALLOWED)
         collection.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
